refactor: put components into separate files
