+++ src/components/search-panel.js
@@ -1,0 +1,5 @@
+const SearchPanel = () => {
+  return <input placeholder="search" />;
+};
+
+export default SearchPanel;
