+++ src/components/todo-list-item.js
@@ -1,0 +1,5 @@
+const TodoListItem = () => {
+  return <span>Drink coffee</span>;
+};
+
+export default TodoListItem;
