+++ src/index.js
@@ -1,20 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
-
-const TodoList = () => {
-  return (
-    <ul>
-      <li>Learn React</li>
-      <li>Build Awesome App</li>
-    </ul>
-  );
-};
-const AppHeader = () => {
-  return <h1>My Todo List</h1>;
-};
-const SearchPanel = () => {
-  return <input placeholder="search" />;
-};
+import AppHeader from "./components/app-header";
+import SearchPanel from "./components/search-panel";
+import TodoList from "./components/todo-list";
 
 const App = () => (
   <div>
